+++ CHANGELOG.md
@@ -17,8 +17,10 @@ Prefix the change with one of these keywords:
 ## Unreleased
 
 - Changed: updated the styling of `Alert` conform to design system
+- Changed: Select can now handle large texts by overflowing into an ellipsis
 - Changed: **BREAKING** removed the `compact` prop/variant in `Alert`
 - Changed: **BREAKING** `svgFill` now only accepts CSS color literal or a `ThemeFn` which returns a CSS color (see: https://github.com/Amsterdam/amsterdam-styled-components/pull/594).
+- Changed: **BREAKING** elements using `aria-hidden` are no longer set to `display: none`
 
 ## [0.20.0]