+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Prefix the change with one of these keywords:
 - Changed: updated the styling of `Alert` conform to design system
 - Changed: **BREAKING** removed the `compact` prop/variant in `Alert`
 - Changed: **BREAKING** `svgFill` now only accepts CSS color literal or a `ThemeFn` which returns a CSS color (see: https://github.com/Amsterdam/amsterdam-styled-components/pull/594).
+- Deprecated: `onDarkBackground` prop in `Link`. Use `darkBackground` instead.
 
 ## [0.20.0]