+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Prefix the change with one of these keywords:
 - Fixed: Button in the SearchBar component is now aligned with the TextField
 - Fixed: Toggle now hides elements inside properly
 - Deprecated: `onDarkBackground` prop in `Link`. Use `darkBackground` instead.
+- Added: make it possible to define which boundaries should be used for edge detection
 
 ## [0.20.0]